requirements.txt
asgiref-3.4.1 
django-3.2.10 
pytz-2021.3 
sqlparse-0.4.2 
typing-extensions-4.0.1